requirement.txt
tencentcloud-sdk-python
progress